+++ internal/dataplane/translator/translate_httproute.go
@@ -3,7 +3,6 @@ package translator
 import (
 	"errors"
 	"fmt"
-	"time"
 
 	"github.com/kong/go-kong/kong"
 	"github.com/samber/lo"
@@ -40,56 +39,8 @@ func (t *Translator) ingressRulesFromHTTPRoutes() ingressRules {
 		httpRoutesToTranslate = append(httpRoutesToTranslate, httproute)
 	}
 
-	if t.featureFlags.ExpressionRoutes {
-		t.ingressRulesFromHTTPRoutesUsingExpressionRoutes(httpRoutesToTranslate, &result)
-		return result
-	}
-
 	t.ingressRulesFromHTTPRoutesWithCombinedService(httpRoutesToTranslate, &result)
 	return result
-}
-
-// applyTimeoutsToService applies timeouts from HTTPRoute to the service.
-// If the HTTPRoute has multiple rules, the timeout from the last rule which has specific timeout will be applied to the service.
-// If the HTTPRoute has multiple rules and the first rule doesn't have timeout, the default timeout will be applied to the service.
-func applyTimeoutsToService(httpRoute *gatewayapi.HTTPRoute, rules *ingressRules) {
-	// If the HTTPRoute doesn't have rules, we don't need to apply timeouts to the service.
-	if httpRoute.Spec.Rules == nil {
-		return
-	}
-
-	backendRequestTimeout := DefaultServiceTimeout
-	for _, rule := range httpRoute.Spec.Rules {
-		if rule.Timeouts != nil && rule.Timeouts.BackendRequest != nil {
-			duration, err := time.ParseDuration(string(*rule.Timeouts.BackendRequest))
-			// We ignore the error here because the rule.Timeouts.BackendRequest is validated
-			// to be a strict subset of Golang time.ParseDuration so it should never happen
-			if err != nil {
-				continue
-			}
-			backendRequestTimeout = int(duration.Milliseconds())
-		}
-	}
-
-	// if the backendRequestTimeout is the same as the default timeout, we don't need to apply it to the service.
-	if backendRequestTimeout == DefaultServiceTimeout {
-		return
-	}
-
-	// due rules.ServiceNameToServices is a map, we need to iterate over the map to find the service
-	// which has the same parent as the HTTPRoute.
-	for serviceName, service := range rules.ServiceNameToServices {
-		if service.Parent.GetObjectKind() == httpRoute.GetObjectKind() && service.Parent.GetName() == httpRoute.Name && service.Parent.GetNamespace() == httpRoute.Namespace {
-			// Due to only one field being available in the Gateway API to control this behavior,
-			// when users set `spec.rules[].timeouts` in HTTPRoute,
-			// KIC will also set ReadTimeout, WriteTimeout and ConnectTimeout for the service to this value
-			// https://github.com/Kong/kubernetes-ingress-controller/issues/4914#issuecomment-1813964669
-			service.Service.ReadTimeout = kong.Int(backendRequestTimeout)
-			service.Service.ConnectTimeout = kong.Int(backendRequestTimeout)
-			service.Service.WriteTimeout = kong.Int(backendRequestTimeout)
-			rules.ServiceNameToServices[serviceName] = service
-		}
-	}
 }
 
 func validateHTTPRoute(httproute *gatewayapi.HTTPRoute, featureFlags FeatureFlags) error {
@@ -122,15 +73,15 @@ func validateHTTPRoute(httproute *gatewayapi.HTTPRoute, featureFlags FeatureFlag
 // When the feature flag CombinedServicesFromDifferentHTTPRoutes is true, it combines rules with same backends
 // to a single Kong gateway service across different HTTPRoutes in the same namespace.
 // When the feature flag is false, it combines rules with same backends in an HTTPRoute to a Kong gateway service.
-// TODO: This would be the common entrance for translating to Kong services in both traditional and expression routes after
-// https://github.com/Kong/kubernetes-ingress-controller/issues/6727 is finished.
-// They would share the code for translation of service names, backends and fields other then routes in Kong services.
+// When the feature flag ExpressionRoutes is set to true, expression based Kong routes will be translated from matches of HTTPRoutes.
+// Otherwise, traditional Kong routes are translated.
 func (t *Translator) ingressRulesFromHTTPRoutesWithCombinedService(httpRoutes []*gatewayapi.HTTPRoute, result *ingressRules) {
 	translationResult := subtranslator.TranslateHTTPRoutesToKongstateServices(
 		t.logger,
 		t.storer,
 		httpRoutes,
 		t.featureFlags.CombinedServicesFromDifferentHTTPRoutes,
+		t.featureFlags.ExpressionRoutes,
 	)
 	for serviceName, service := range translationResult.ServiceNameToKongstateService {
 		result.ServiceNameToServices[serviceName] = service
@@ -151,52 +102,6 @@ func (t *Translator) ingressRulesFromHTTPRoutesWithCombinedService(httpRoutes []
 			continue
 		}
 		// Register HTTPRoute successfully translated if no translation error found.
-		t.registerSuccessfullyTranslatedObject(httproute)
-	}
-}
-
-// ingressRulesFromHTTPRoutesUsingExpressionRoutes translates HTTPRoutes to expression based routes
-// when ExpressionRoutes feature flag is enabled.
-// Because we need to assign different priorities based on the hostname and match in the specification of HTTPRoutes,
-// We need to split the HTTPRoutes into ones with only one hostname and one match, then assign priority to them
-// and finally translate the split HTTPRoutes into Kong services and routes with assigned priorities.
-func (t *Translator) ingressRulesFromHTTPRoutesUsingExpressionRoutes(httpRoutes []*gatewayapi.HTTPRoute, result *ingressRules) {
-	// first, split HTTPRoutes by hostnames and matches.
-	splitHTTPRouteMatches := []subtranslator.SplitHTTPRouteMatch{}
-	for _, httproute := range httpRoutes {
-		splitHTTPRouteMatches = append(splitHTTPRouteMatches, subtranslator.SplitHTTPRoute(httproute)...)
-	}
-	// assign priorities to split HTTPRoutes.
-	splitHTTPRoutesWithPriorities := subtranslator.AssignRoutePriorityToSplitHTTPRouteMatches(t.logger, splitHTTPRouteMatches)
-	httpRouteNameToTranslationFailure := map[k8stypes.NamespacedName][]error{}
-
-	// translate split HTTPRoute matches to ingress rules, including services, routes, upstreams.
-	for _, httpRouteWithPriority := range splitHTTPRoutesWithPriorities {
-		err := t.ingressRulesFromSplitHTTPRouteMatchWithPriority(result, httpRouteWithPriority)
-		if err != nil {
-			nsName := k8stypes.NamespacedName{
-				Namespace: httpRouteWithPriority.Match.Source.Namespace,
-				Name:      httpRouteWithPriority.Match.Source.Name,
-			}
-			httpRouteNameToTranslationFailure[nsName] = append(httpRouteNameToTranslationFailure[nsName], err)
-		}
-	}
-	// Register successful translated objects and translation failures.
-	// Because one HTTPRoute may be split into multiple HTTPRoutes, we need to de-duplicate by namespace and name.
-	for _, httproute := range httpRoutes {
-		nsName := k8stypes.NamespacedName{
-			Namespace: httproute.Namespace,
-			Name:      httproute.Name,
-		}
-		if translationFailures, ok := httpRouteNameToTranslationFailure[nsName]; !ok {
-			applyTimeoutsToService(httproute, result)
-		} else {
-			t.registerTranslationFailure(
-				fmt.Sprintf("HTTPRoute can't be routed: %v", errors.Join(translationFailures...)),
-				httproute,
-			)
-			continue
-		}
 		t.registerSuccessfullyTranslatedObject(httproute)
 	}
 }
@@ -248,7 +153,6 @@ func GenerateKongRouteFromTranslation(
 
 	// get the hostnames from the HTTPRoute
 	hostnames := getHTTPRouteHostnamesAsSliceOfStringPointers(httproute)
-
 	return subtranslator.GenerateKongRoutesFromHTTPRouteMatches(
 		translation.Name,
 		translation.Matches,
@@ -257,59 +161,4 @@ func GenerateKongRouteFromTranslation(
 		hostnames,
 		tags,
 	)
-}
-
-func httpBackendRefsToBackendRefs(httpBackendRef []gatewayapi.HTTPBackendRef) []gatewayapi.BackendRef {
-	backendRefs := make([]gatewayapi.BackendRef, 0, len(httpBackendRef))
-
-	for _, hRef := range httpBackendRef {
-		backendRefs = append(backendRefs, hRef.BackendRef)
-	}
-	return backendRefs
-}
-
-// ingressRulesFromSplitHTTPRouteMatchWithPriority translates a single match split from HTTPRoute
-// to ingress rule, including Kong service and Kong route.
-func (t *Translator) ingressRulesFromSplitHTTPRouteMatchWithPriority(
-	rules *ingressRules,
-	httpRouteMatchWithPriority subtranslator.SplitHTTPRouteMatchToKongRoutePriority,
-) error {
-	match := httpRouteMatchWithPriority.Match
-	httpRoute := httpRouteMatchWithPriority.Match.Source
-	if match.RuleIndex >= len(httpRoute.Spec.Rules) {
-		t.logger.Error(nil, "Split match has rule out of bound of rules in source HTTPRoute",
-			"rule_index", match.RuleIndex, "rule_count", len(httpRoute.Spec.Rules))
-		return nil
-	}
-
-	rule := httpRoute.Spec.Rules[match.RuleIndex]
-	backendRefs := httpBackendRefsToBackendRefs(rule.BackendRefs)
-	serviceName := subtranslator.KongServiceNameFromSplitHTTPRouteMatch(httpRouteMatchWithPriority.Match)
-
-	kongService, err := generateKongServiceFromBackendRefWithName(
-		t.logger,
-		t.storer,
-		rules,
-		serviceName,
-		httpRoute,
-		"http",
-		backendRefs...,
-	)
-	if err != nil {
-		return err
-	}
-
-	additionalRoutes, err := subtranslator.KongExpressionRouteFromHTTPRouteMatchWithPriority(httpRouteMatchWithPriority)
-	if err != nil {
-		return err
-	}
-
-	kongService.Routes = append(
-		kongService.Routes,
-		*additionalRoutes,
-	)
-	// cache the service to avoid duplicates in further loop iterations
-	rules.ServiceNameToServices[serviceName] = kongService
-	rules.ServiceNameToParent[serviceName] = httpRoute
-	return nil
 }