+++ internal/manager/controllerdef.go
@@ -154,17 +154,7 @@ func setupControllers(
 				DataplaneClient:   dataplaneClient,
 				CacheSyncTimeout:  c.CacheSyncTimeout,
 				ReferenceIndexers: referenceIndexers,
-			},
-		},
-		{
-			Enabled: true,
-			Controller: &configuration.CoreV1ConfigMapReconciler{
-				Client:            mgr.GetClient(),
-				Log:               ctrl.LoggerFrom(ctx).WithName("controllers").WithName("configmaps"),
-				Scheme:            mgr.GetScheme(),
-				DataplaneClient:   dataplaneClient,
-				CacheSyncTimeout:  c.CacheSyncTimeout,
-				ReferenceIndexers: referenceIndexers,
+				LabelSelector:     c.ConfigMapLabelSelector,
 			},
 		},
 		// ---------------------------------------------------------------------------